[FEAT] 프로필, 지갑 기본기능 구현

+++ src/main/java/io/nh_backend/rest_quest/user/dto/ShowWalletResponse.java
@@ -1,7 +1,7 @@
 package io.nh_backend.rest_quest.user.dto;
 
 public record ShowWalletResponse(
-        Integer gold,
-        Integer gem
+        Long gold,
+        Long gem
 ) {
 }
