+++ src/main/java/io/nh_backend/rest_quest/user/dto/UserProfileResponse.java
@@ -2,6 +2,6 @@ package io.nh_backend.rest_quest.user.dto;
 
 public record UserProfileResponse(
         Integer level,
-        Integer exp
+        Long exp
 ) {
 }
